게시글 목록조회 ,작성 게시글 조회 API ->페이징 기능 제거

+++ Server/src/main/java/ga/healtCare/src/cleaningInfo/model/CleaningInfo.java
@@ -30,7 +30,7 @@ public class CleaningInfo extends BaseEntity {
     private Integer liquidAmount;
 
     @Column(name = "recentlyCleaningTime")
-    private Integer recentlyCleaningTime;
+    private Long recentlyCleaningTime;
 
     @Column(name = "isCleaned")
     private Integer isCleaned;
